# Dumping response body (optimized multi-writer)