--- a/fox_test.go
+++ b/fox_test.go
@@ -480,6 +480,26 @@ func benchRouteParallel(b *testing.B, router http.Handler, rte route) {
 	})
 }
 
+/*func BenchmarkDumpBodyFox(b *testing.B) {
+	buf := bytes.NewBuffer(nil)
+	f := New()
+	f.MustHandle(http.MethodGet, "/hello/{name}", func(c Context) {
+		buf.Reset()
+		c.TeeWriter(buf)
+		_, _ = io.WriteString(c.Writer(), c.Param("name"))
+	})
+
+	w := new(mockResponseWriter)
+	r := httptest.NewRequest("GET", "/hello/fox", nil)
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		f.ServeHTTP(w, r)
+	}
+}*/
+
 func BenchmarkStaticAll(b *testing.B) {
 	r := New()
 	for _, route := range staticRoutes {