### Abstract
This proposal introduces a new feature in the Fox Router to enable users to dump the HTTP response body while writing to the `ResponseWriter`. The feature is designed to be efficient, versatile, and compatible with existing `http.ResponseWriter` interfaces like `http.Hijacker`, `http.Flusher,` etc. To achieve this, the proposal introduces a new API method and constructs custom multiwriters compatible with HTTP/1 and HTTP/2 interface.

### Introduction
In the current system, there is no direct way to dump the HTTP response body while writing to the `ResponseWriter`. This feature can be useful in several scenarios, such as debugging, logging, testing, and monitoring. This proposal aims to fill this gap.

### Proposal
The proposed change introduces a new API method:
````go
type Context inteface {
    // TeeWriter set up one or more additional writers (sink) to which the response body will be written.
    TeeWriter(sink io.Writer,sinks ...io.Writer) {...}
}
````

When invoked, this method sets up one or more additional writers (sinks) to which the response body will be written. These writers are typically `io.Writer` implementations like `bytes.Buffer`. Depending on the HTTP protocol in use, the router creates either an `h1MultiWriter` or an `h2MultiWriter` that ensures compatibility with existing interfaces like `ReaderFrom` or `Pusher`.

The `h1MultiWriter` and `h2MultiWriter` are custom implementations of  `io.MultiWriter`. They handle writing to multiple writers, including the original `ResponseWriter` and any additional writers specified by the user. When a write operation occurs, it writes to all these writers. Similarly, when a flush operation is invoked, it flushes all writers that implement the `http.Flusher` interface. For HTTP/2, it also handles server push operations.

This design ensures that the new feature works efficiently and seamlessly with the existing system, without breaking any current functionality or compatibility.

### Performance Considerations
This feature is designed with performance in mind. When the `TeeWriter` feature is not used, there is no performance impact on the router's operation. The creation of the custom multiwriters only occurs when the `TeeWriter` method is invoked with one or more writers. This ensures that the overhead of managing multiple writers is only incurred when the feature is actively used.

For example, the following preliminary benchmark offers a comparison between Echo's `ResponseBodyDumper` middleware and the proposed `TeeWriter` method.

````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
BenchmarkDumpBodyFox-16          8429208               139.8 ns/op            16 B/op          1 allocs/op
BenchmarkDumpBodyEcho-16         4582627               308.4 ns/op           128 B/op          5 allocs/op
````

### Flexibility and Use Cases
The `TeeWriter` method is extremely flexible and can be used in various ways to suit different use cases. It can be used in a middleware, providing a consistent body dumping behavior across multiple routes, or it can be used on a per-route basis, allowing for fine-grained control over which routes have the body dumping feature enabled.

Furthermore, because the user provides the `io.Writer` to the `TeeWriter` method, they have complete control over the management of these writers. For instance, they could use a `sync.Pool` of writers to efficiently reuse memory across multiple requests, or they could use different writer implementations for different routes or conditions. This allows users to tailor the body dumping behavior to their specific needs and constraints.

### Backward Compatibility
The proposed change is fully backward compatible. It introduces a new API method and does not modify or remove any existing APIs. Users who do not use the new feature will not experience any change in the router's behavior.


### Conclusion
The proposed feature enhances the router functionality by allowing users to dump the HTTP response body while writing to the `ResponseWriter`. It is designed to be efficient, flexible, and reasonably compatible with existing `http.ResponseWriter` interfaces. The feature can be used as needed, without any performance impact when not in use. With the user's freedom to provide their own `io.Writer`, they can implement the feature in a way that best suits their specific use cases, whether it be efficient memory usage with `sync.Pool`, or different writer implementations for different routes or conditions. This feature will be a valuable tool for debugging, testing, logging, and monitoring HTTP response data.