@tigerwill90 2023-06-06
````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
                    │   old.txt    │               new.txt                │
                    │    sec/op    │    sec/op      vs base               │
StaticAll-16          11.01µ ±  1%    11.27µ ±  2%  +2.36% (p=0.000 n=10)
GithubParamsAll-16    82.67n ±  1%    87.49n ±  2%  +5.83% (p=0.000 n=10)
OverlappingRoute-16   98.34n ±  1%   100.35n ±  1%  +2.05% (p=0.001 n=10)
StaticParallel-16     9.898n ± 10%    9.216n ±  2%  -6.90% (p=0.000 n=10)
CatchAll-16           32.57n ±  3%    32.77n ±  1%       ~ (p=0.197 n=10)
CatchAllParallel-16   4.949n ± 21%    5.487n ± 19%       ~ (p=0.089 n=10)
PathClean-16          790.8n ±  1%    772.2n ±  1%  -2.34% (p=0.000 n=10)
PathCleanLong-16      4.482m ± 18%    4.237m ± 15%       ~ (p=0.481 n=10)
MultiWriter-16                        101.8n ±  2%
geomean               387.3n          336.0n        +0.73%
````
The benchmark results indicate that there are no significant performance impacts in the new implementation compared to the previous one.